requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
Flask-Cors
python-dotenv
requests
requests==2.31.0
urllib3==2.2.1
Google-Images-Search==1.4.6
beautifulsoup4==4.10.0
python-dotenv==0.19.2
openpyxl